moved cantabular client interface to api package
