+++ service/interfaces.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"context"
-	"github.com/ONSdigital/dp-dataset-api/cantabular"
+	"github.com/ONSdigital/dp-dataset-api/api"
 	"net/http"
 
 	"github.com/ONSdigital/dp-dataset-api/config"
@@ -23,7 +23,7 @@ type Initialiser interface {
 	DoGetKafkaProducer(ctx context.Context, cfg *config.Configuration, topic string) (kafka.IProducer, error)
 	DoGetGraphDB(ctx context.Context) (store.GraphDB, Closer, error)
 	DoGetMongoDB(ctx context.Context, cfg config.MongoConfig) (store.MongoDB, error)
-	DoGetCantabular(ctx context.Context, cfg config.CantabularConfig) cantabular.CantabularClient
+	DoGetCantabular(ctx context.Context, cfg config.CantabularConfig) api.CantabularClient
 }
 
 // HTTPServer defines the required methods from the HTTP server
