+++ service/service.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"context"
-	"github.com/ONSdigital/dp-dataset-api/cantabular"
 	"net/http"
 
 	clientsidentity "github.com/ONSdigital/dp-api-clients-go/v2/identity"
@@ -38,7 +37,7 @@ type DatsetAPIStore struct {
 type Service struct {
 	config                              *config.Configuration
 	serviceList                         *ExternalServiceList
-	cantabular                          cantabular.CantabularClient
+	cantabular                          api.CantabularClient
 	graphDB                             store.GraphDB
 	graphDBErrorConsumer                Closer
 	mongoDB                             store.MongoDB
